Fixed OBJ, parser now it supports multiple objects.. Still to work on texture and materials.

+++ TrikytaEngine3D/Core/Misc/Utils/Common.hpp
@@ -10,8 +10,8 @@ typedef std::string String;
 template<typename K, typename T>
 using HashMap = std::map<K, T>;
 
-template <uint64 N>
-static FORCEINLINE uint64 GetArraySize(float(&p_Array)[N]) {
-	uint64 num = (sizeof(p_Array) / sizeof(p_Array[0]));
+template <usize N>
+static FORCEINLINE usize GetArraySize(float(&p_Array)[N]) {
+	usize num = (sizeof(p_Array) / sizeof(p_Array[0]));
 	return num;
 }
